+++ src/example/app/main/module.py
@@ -1,7 +1,7 @@
 from dependency.core import Module, module
-from example.plugins.base import BaseModule
-from example.plugins.hardware import HardwareModule
-from example.plugins.reporter import ReporterModule
+from example.plugin.base import BaseModule
+from example.plugin.hardware import HardwareModule
+from example.plugin.reporter import ReporterModule
 
 @module(
     imports=[
