+++ src/example/plugin/hardware/bridge/__init__.py
