+++ src/example/plugin/hardware/factory/interfaces.py
