+++ src/example/plugin/hardware/factory/products/productA.py
@@ -1,0 +1,17 @@
+from dependency.core import dependent
+from dependency.core.declaration.dependent import Dependent
+from example.plugin.hardware.factory.interfaces import Hardware
+from example.plugin.base.number import NumberService, NumberServiceComponent
+
+@dependent(
+    imports=[
+        NumberServiceComponent
+    ],
+)
+class HardwareA(Hardware, Dependent):
+    def __init__(self) -> None:
+        self.__number: NumberService = NumberServiceComponent.provide()
+
+    def doStuff(self, operation: str) -> None:
+        random_number = self.__number.getRandomNumber()
+        print(f"HardwareA {random_number} works with operation: {operation}")
