+++ src/example/plugin/hardware/observer/__init__.py
@@ -1,0 +1,19 @@
+from abc import ABC, abstractmethod
+from dependency.core import Component, component
+from typing import Callable
+from example.plugin.hardware.observer.interfaces import HardwareEventContext, EventSubscriber
+
+class HardwareObserver(ABC):
+    @abstractmethod
+    def subscribe(self, listener: type[EventSubscriber]) -> Callable:
+        pass
+
+    @abstractmethod
+    def update(self, context: HardwareEventContext) -> None:
+        pass
+
+@component(
+    interface=HardwareObserver
+)
+class HardwareObserverComponent(Component):
+    pass
