+++ src/example/plugin/reporter/factory/__init__.py
@@ -1,0 +1,14 @@
+from abc import ABC, abstractmethod
+from dependency.core import Component, component
+from example.plugin.reporter.factory.interfaces import Reporter
+
+class ReporterFactory(ABC):
+    @abstractmethod
+    def createProduct(self, product: str) -> Reporter:
+        pass
+
+@component(
+    interface=ReporterFactory
+)
+class ReporterFactoryComponent(Component):
+    pass
