--- src/example/plugins/base/number/providers/fake.py
@@ -1,13 +1,0 @@
-from dependency.core import provider, providers
-from example.plugins.base.number import NumberService, NumberServiceComponent
-
-@provider(
-    component=NumberServiceComponent,
-    provider = providers.Singleton
-)
-class FakeNumberService(NumberService):
-    def __init__(self, config: dict) -> None:
-        pass
-
-    def getRandomNumber(self) -> int:
-        return 42
