--- src/example/plugins/hardware/factory/products/productC.py
@@ -1,8 +1,0 @@
-from dependency.core import dependent
-from dependency.core.declaration.dependent import Dependent
-from example.plugins.hardware.factory.interfaces import Hardware
-
-@dependent()
-class HardwareC(Hardware, Dependent):
-    def doStuff(self, operation: str) -> None:
-        print(f"HardwareC works with operation: {operation}")
